% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_financial_df.R
\name{construct_financial_df}
\alias{construct_financial_df}
\title{Construct Attributes for a Financial Data Frame}
\usage{
construct_financial_df(df, crypto = F, ...)
}
\arguments{
\item{df}{\code{data.frame} The function compute statistics with the \code{to} column (if available) or the \code{currency} column}

\item{crypto}{\code{logical} Attribute predefined if cryptocurrencies are in the dataset}
}
\description{
This internal function constructs and assigns specific attributes
to a financial data frame, providing metadata about its contents.
}
\details{
Return a \code{data.frame} with additionnal attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
}
}
\seealso{
Other functions that use these attributes can reference this documentation.
}
\keyword{internal}
