% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.R
\name{get_asset_value}
\alias{get_asset_value}
\title{Retrieve flat meta data for a financial asset}
\usage{
get_asset_value(symbol = "SAAB-B.ST", .verbose = F)
}
\arguments{
\item{symbol}{Character. Asset symbol to query (e.g., "AAPL").}

\item{.verbose}{Logical. If TRUE, enables verbose output for debugging purposes. Default is FALSE.}
}
\value{
A data.frame with 1 row per symbol provided and 21 columns see \code{get_values()}
}
\description{
Internal function to fetch flat (non-nested) meta data for a given asset symbol, such as a stock.
Returns a one-row data frame with high-level market and asset information (e.g., prices, volumes, timezones).
}
\examples{
\dontrun{
last_prices <- get_asset_value("SAAB-B.ST")
str(last_prices)
}
}
\seealso{
\code{get_values}
}
\keyword{internal}
