% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_changes.R
\name{get_changes}
\alias{get_changes}
\title{Get Latest Exchange Rates For Devises}
\usage{
get_changes(from = NULL, to = "USD", .verbose = T, ...)
}
\arguments{
\item{from, }{default = \code{NULL} A character string representing the base currency (e.g., "USD").
The user must provide \code{from} values with one of the following method :
\itemize{
\item passing a paired character list, e.g., c("EUR" = "USD", "RON" = "EUR") ;
\item or passing one or several \code{from} values :
\itemize{
\item associated with a single \code{to} value, e.g., \verb{from = c('EUR', "RON"), to = "USD"}
\item or associated with a list of \code{from} values of the same length, e.g., \verb{from = c('EUR', "RON"), to = c("USD", "EUR")} .
}
}}

\item{to, }{default = \code{"USD"} A character string representing the target currencies, e.g., \code{c("EUR", "USD")}.}

\item{.verbose}{\code{logical} If TRUE, messages are displayed, e.g., when invalid symbols are detected. Default is FALSE.}

\item{...}{Other parameters passed to get_a_change - internal mecanism for get_changes_historic}
}
\value{
The returned dataframe contains daily exchange rates, and the following columns:
\describe{
\item{currency}{\code{character}, the base currency.}
\item{symbol}{\code{character}, the Yahoo Finance symbol (e.g., "EURUSD=X").}
\item{exchange_name}{\code{character}, the exchange place name, i.e. 'CCY' for currencies.}
\item{fullexchange_name}{\code{character}, the full exchange place name, supposed to be 'CCY' for currencies.}
\item{intrumenttype}{\code{character}, the type of financial instrument, supposed to be 'CURRENCY'.}
\item{firsttradedate}{\code{POSIXct}, the oldest date available on the Yahoo Finance API for an exchange rates historic.}
\item{regularmarketprice}{\code{numeric}, the latest market price.}
\item{regularmarketdayhigh}{\code{numeric}, the market highest price of this day.}
\item{regularmarketdaylow}{\code{numeric}, the market lowest price of this day.}
\item{fifty_two_week_high}{\code{numeric}, the highest price in the last 52 weeks.}
\item{fifty_two_week_low}{\code{numeric}, the lowest price in the last 52 weeks.}
\item{previous_close}{\code{numeric}, the last closing price.}
\item{timezone}{\code{character}, the market's timezone.}
\item{date}{\code{POSIXct} Date of the observation.}
\item{from}{\code{character}, the currency converted into another, e.g., if the \code{from} value is 1$ ('USD'), you want to receive a certain amount of the other currency to reach 1$.}
\item{to}{\code{character}, the currency that you want to convert into : \strong{all the \code{numeric} values (not \code{integer}) in this line of the \code{data.frame} are expressed with this currency}.}
\item{timestamp}{int, the timestamp of the rate.}
\item{rate.high}{Numeric, the highest exchange rate of the day.}
\item{rate.low}{Numeric, the lowest exchange rate of the day.}
\item{rate.open}{Numeric, the opening exchange rate of the day.}
\item{rate.volume}{Numeric, the trading volume for the day.}
\item{rate.close}{Numeric, the closing exchange rate of the day.}
\item{timestamp}{\code{POSIXct}, the corresponding date (YYYY-MM-DD).}
}
}
\description{
Returns a \code{data.frame} with the latest exchange rate(s) for the given currencies. Default will convert to USD ($).
}
\details{
Return a \code{data.frame} with additionnal attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
}
}
\examples{
# Fetch exchange historical data from € and ¥ to $ (default convert to 'USD")
df <- get_changes(from = c("EUR", "JPY"))
str(df)
# or pass a named list of character
df2 <- get_changes(from = c("EUR" = "RON", "USD" = "EUR"))
# Or pass paired values as 2 list (equivalent to hereabove line) :
same_as_df2 <- get_changes(from = c("EUR", "USD"), to =c("RON" , "EUR"))
}
\references{
Source : https://query1.finance.yahoo.com/v8/finance/chart/
}
