% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.R
\name{get_changes_historic}
\alias{get_changes_historic}
\title{Get Historic of Devises Exchanges Rates}
\usage{
get_changes_historic(
  from = NULL,
  to = "USD",
  interval = "1d",
  range = "1y",
  .verbose = T
)
}
\arguments{
\item{from}{default = \code{NULL} A character string representing the base currency (e.g., "USD").
The user must provide \code{from} values with one of the following method :
\itemize{
\item passing a paired character list, e.g., c("EUR" = "USD", "RON" = "EUR") ;
\item or passing one or several \code{from} values :
\itemize{
\item associated with a single \code{to} value, e.g., \verb{from = c('EUR', "RON"), to = "USD"}
\item or associated with a list of \code{from} values of the same length, e.g., \verb{from = c('EUR', "RON"), to = c("USD", "EUR")} .
}
}}

\item{to}{default = \code{"USD"} A character string representing the target currencies, e.g., \code{c("EUR", "USD")}.}

\item{interval}{\code{character}, default = \code{"1d"}. The interval between 2 rows of the time.series answered\tabular{l}{
   \strong{Valid \code{interval} values} \cr
   \code{"1m"}, \code{"2m"}, \code{"3m"}, \code{"5m"}, \code{"15m"}, \code{"30m"}, \code{"1h"}, \code{"4h"}, \code{"1d"}, \code{"1wk"}, \code{"1mo"}, \code{"3mo"} \cr
}}

\item{range}{\code{character}, default = \code{"1y"}. The period covered by the time series.\tabular{l}{
   \strong{Valid \code{range} values} \cr
   \code{"1d"}, \code{"5d"}, \code{"1m"}, \code{"3m"}, \code{"6m"}, \code{"1y"}, \code{"5y"}, \code{"ytd"}, \code{"all"} \cr
}}

\item{.verbose}{\code{logical} If TRUE, messages are displayed, e.g., when invalid symbols are detected. Default is FALSE.}
}
\value{
A \code{data.frame} with the historic of exchanges rates on a given period, default is daily results for a year.
The columns are:
\item{timestamp}{\code{POSIXct} - The opening price for the period (default is each day).}
\item{close}{\code{numeric} - The closing price for the period (default is each day).}
\item{low}{\code{numeric} - The highest price for the period (default is each day).}
\item{open}{\code{integer} - The traded volume.}
\item{high}{\code{numeric} - The lowest price for the period (default is each day).}
\item{from}{\code{character} - The currency converted into another, e.g., if the \code{from} value is 1$ ('USD'), you want to receive a certain amount of the other (\code{to}) currency to reach 1$.}
\item{to}{\code{character} -  The currency exchanged back against a value of 1 of the \code{from} currency. The currencies-related \code{numeric} values in this line of the \code{data.frame} are expressed with this currency**.}
Depending on the desired interval, recent observation will be truncated,
e.g., a \code{'5y'} range  with a \code{'1d'} interval will answer approximately 30 days of values from 5 years ago.
}
\description{
Get a \code{data.frame} of historical values of exchanges rates, given currencies to exchange.
Default parameters are a period of 1 year (1 obs. per day), and convert to USD ($).
}
\details{
Return a \code{data.frame} with additional attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date.fetch}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
\item{date.begin}{\code{POSIXct} - The oldest observation in the dataset.}
\item{date.end}{\code{POSIXct} - The most recent observation in the dataset.}
\item{date.dif}{\code{numeric} - The number of \emph{seconds} between date.begin and date.end, equivalent of \code{difftime} value.}
}
See \code{vignette("Functions_summary", package = "financr")}
}
\examples{
convert_to_usd <- get_changes_historic(from = c("EUR", "JPY"))

days_bis <- get_changes_historic(from = c("EUR" = "RON", "USD" = "EUR"))

# (equivalent to the previous line)
# same_as_days_bis <- get_changes_historic(from = c("EUR", "USD"), to =c("RON" , "EUR"))

# Tweak interval and range
months <- get_changes_historic(from = c("EUR", "JPY"), interval = "1mo", range = '5y')
str(months)
}
\references{
Source : https://query2.finance.yahoo.com/v8/finance/chart/
}
\seealso{
\code{\link{get_changes}}

For more details see the help vignette:
\code{vignette("currencies", package = "financr")}
}
