% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_currencies.R
\name{get_currencies}
\alias{get_currencies}
\title{Get Currencies Names And Symbol}
\usage{
get_currencies(keep = NULL, .verbose = T)
}
\arguments{
\item{keep}{A character vector of symbols to filter the results. If NULL (default),
no filtering is applied, and all available indices are returned.}

\item{.verbose}{\code{logical}, default = \code{TRUE}. If \code{TRUE}, send messages to the console if necessary.}
}
\value{
A data frame containing unique financial indices (currencies). The table has
columns like \code{symbol}, \code{name}, and other relevant information, with all column names in lowercase.
If \code{keep} is specified, only the matching indices are returned.#'
}
\description{
Return a \code{data.frame} of financial indices (currencies) from Yahoo Finance.
Optionally, it can filter the results, given a vector of symbols, e.g., 'EUR'.
}
\examples{
# Fetch all available indices
all_indices <- get_currencies()

# Fetch only specific indices
selected_indices <- get_currencies(keep = c("USD", "EUR"))
}
\references{
Source : https://query1.finance.yahoo.com/v1/finance/currencies
}
