% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historic.R
\name{get_historic}
\alias{get_historic}
\title{Get Historical Financial Data Indices}
\usage{
get_historic(symbols = c("SAAB-B.ST"), wait.time = 0, .verbose = T, ...)
}
\arguments{
\item{symbols}{\code{character} A character string representing the financial indices to search for, e.g., ticker symbol(s).}

\item{wait.time}{\code{double}, default = \code{0} A character string representing an additional waiting time between 2 calls to the Yahoo API.}

\item{.verbose}{\code{logical}, default = \code{TRUE}. If \code{TRUE}, send messages to the console.}

\item{...}{
  Arguments passed on to \code{\link[=get_yahoo_data]{get_yahoo_data}}
  \describe{
    \item{\code{symbol}}{\code{character} A character string representing the symbol for the financial instrument (e.g., "AAPL" for Apple).}
    \item{\code{start_date}}{\code{character} A character string representing the start date in the format "YYYY-MM-DD". If \code{NULL}, data starts from 1970-01-01.}
    \item{\code{end_date}}{\code{character} A character string representing the end date in the format "YYYY-MM-DD". If \code{NULL}, data is retrieved up to the current date.}
    \item{\code{range}}{\code{character} A character string representing the range for the returned datas : default will return daily values.\tabular{l}{
   \strong{Valid \code{range} values} \cr
   \code{"1d"}, \code{"5d"}, \code{"1mo"}, \code{"3mo"}, \code{"6mo"}, \code{"1y"}, \code{"2y"}, \code{"5y"}, \code{"10y"}, \code{"ytd"} and \code{"max"} \cr
   Other ranges will filter out some of these daily values, depending on the desired range. \cr
   Valid ranges are "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y", "ytd", and "max". \cr
}}
  }}
}
\value{
A data frame containing the historical financial data with the following columns:
\item{open}{\code{numeric} The opening price for the period (default is each day).}
\item{close}{\code{numeric} The closing price for the period (default is each day).}
\item{low}{\code{numeric} The lowest price for the period (default is each day).}
\item{high}{\code{numeric} The highest price for the period (default is each day).}
\item{volume}{\code{integer} The traded volume.}
\item{timestamp}{\code{integer} Unix timestamps corresponding to each data point.}
\item{date}{\code{POSIXct} The day of the financial data point.}
\item{currency}{\code{character} The currency in which the data is reported, depending on the marketplace.}
\item{symbol}{\code{character} The stock or financial instrument symbol (e.g., "AAPL").}
\item{shortname}{\code{character} The abbreviated name of the company or financial instrument.}
\item{longname}{\code{character} The full name of the company or financial instrument.}
\item{exchangename}{\code{character} The name of the exchange marketplace where the financial instrument is listed.}
\item{fullexchangename}{\code{character} The full name of the exchange marketplace.}
\item{timezone}{\code{character} The timezone in which the data is reported.}
}
\description{
Get the historic of stock market data for financial ticker symbols, e.g., values at each closing day.
}
\details{
Return a \code{data.frame} with additionnal attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
}
}
\examples{
datas <- get_historic(symbols = c("VOLCAR-B.ST", "SAAB-B.ST") )
}
\references{
Source : https://query1.finance.yahoo.com/v8/finance/chart/
}
\seealso{
\code{\link{get_yahoo_data}}
}
