% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_from_name.R
\name{get_info_from_name}
\alias{get_info_from_name}
\title{Get Latest Financial Information and Ticker Symbol From Company Name(s)}
\usage{
get_info_from_name(names, exchange = NULL, sector = NULL)
}
\arguments{
\item{names}{A character string representing the company name to search for.}

\item{exchange}{(optionnal) A character string representing the exchange place(s) to consider (exact match). Default keep all the exchange places.}

\item{sector}{(optionnal)  A character string representing the sector(s) to consider (exact match). Default keep all results.}
}
\value{
A data frame with columns:
- \code{symbol}: The stock ticker symbol from yahoo
- \code{name}: The full company name.
- \code{last_price}: The latest available price.
- \code{sector}: The sector or industry category (if available).
- \code{type}: The type of asset (certainly "stocks").
- \code{exchange}: The stock exchange place for this stock.
\itemize{
\item \code{searched}: The original names searched
}
}
\description{
Get latest overall insights and ticker symbols associated with free texts, such as companies names.
Several ticker symbols associated with various exchanges places could be returned, with
, companies names, last price on the marketplace,
sector/category (if available), etc.
}
\details{
Return a \code{data.frame} with additionnal attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
}
}
\examples{
oil <- get_info_from_name(names = c("TOTAL", "SHELL", "BP"), sector = "Energy")
#Get data on marketplace(s)
swedish <- get_info_from_name(names = c("SAAB", "VOLVO"),  exchange = c("STO", "PAR"))
}
\references{
Source : search on https://finance.yahoo.com/lookup/
}
\seealso{
\code{\link{get_yahoo_data}},  \code{\link{get_historic}}
}
