% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.R
\name{get_similar}
\alias{get_similar}
\title{Find Similar Financial Symbols}
\usage{
get_similar(symbols = "SAAB-B.ST", .verbose = F, ...)
}
\arguments{
\item{symbols}{\code{character} A character string representing a financial symbol to search.}

\item{.verbose}{\code{logical} If TRUE, messages are displayed when invalid symbols are detected. Default is TRUE.}

\item{...}{Other symbols (char. or list of char.)}
}
\value{
A \code{data.frame} with the symbols associated with those provided by the user and similarity scores, according to Yahoo Finance.
\describe{
\item{from}{\code{character} - Financial symbol provided by the user.}
\item{symbol}{\code{character} - Symbol associated with the other 'from' symbol.}
\item{score}{\code{numeric} - Similarity score, according to Yahoo Finance.}
}
}
\description{
Given symbol(s), retrieve identical symbols (according to Yahoo Finance) and a score of similarity.
}
\examples{
get_similar(symbols =   "AAPL,GOOGL")
get_similar(symbols =   c("AAPL", "GOOGL"))
}
\references{
Source : https://query2.finance.yahoo.com/v6/finance/quote/validate
}
