% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.R
\name{get_values}
\alias{get_values}
\title{Retrieve latest financial data}
\usage{
get_values(symbols = c("AAPL", "GOOGL"), .verbose = F)
}
\arguments{
\item{symbols}{Character vector. One or more asset symbols (e.g., \code{c("AAPL", "GOOGL")}).}

\item{.verbose}{Logical. If TRUE, displays verbose output during the fetching process. Default is FALSE.}
}
\value{
A data.frame with 1 row per valid financial symbol and 21 columns, including:
\describe{
\item{currency}{\code{character} - Trading currency of the asset (e.g., "USD").}
\item{symbol}{ \code{character} - Symbol of the asset, e.g., ticker symbol.}
\item{exchangename}{ \code{character} - Abbreviated exchange place or financial instrument name.}
\item{fullexchangename}{ \code{character} - Full name of the exchange place or financial instrument.}
\item{instrumenttype}{ \code{character} - Type of financial instrument (e.g., "EQUITY").}
\item{firsttradedate}{ \code{POSIXct} - Datetime when the asset was first traded.}
\item{regularmarkettime}{ \code{POSIXct} - Timestamp of the latest regular market quote.}
\item{hasprepostmarketdata}{ \code{logical} - Indicates if pre/post-market data is available.}
\item{gmtoffset}{ \code{integer} - Offset from GMT in seconds.}
\item{timezone}{ \code{character} - Abbreviated timezone name of the market.}
\item{exchangetimezonename}{ \code{character} - Timezone name of the exchange location.}
\item{regularmarketprice}{ \code{numeric}   - Latest regular market trading price.}
\item{fiftytwoweekhigh}{ \code{numeric}   - highest price over a 52-weeks period.}
\item{fiftytwoweeklow}{ \code{numeric}   - lowest price over a  52-weeks period.}
\item{regularmarketdayhigh}{ \code{numeric}   - Highest price during the current market day.}
\item{regularmarketdaylow}{ \code{numeric}   - Lowest price during the current market day.}
\item{regularmarketvolume}{ \code{integer}   - Volume traded during the current market day.}
\item{shortname}{ \code{character} - Shortened or common name of the asset.}
\item{chartpreviousclose}{ \code{numeric}   - Closing price shown in charts.}
\item{previousclose}{\code{numeric}   - Previous official market close price.}
}
}
\description{
Get latest insights, given financial symbols. Data such as latest prices and trading context are returned.
}
\details{
Return a \code{data.frame} with additionnal attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date.fetch}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
\item{date.begin}{\code{POSIXct} - The oldest obs. in the dataset.}
\item{date.end}{\code{POSIXct} - The most recent date in the dataset.}
\item{date.dif}{\code{numeric} & \code{difftime} - The number of seconds between date.begin and date.end.}
}
}
\examples{
get_values(c("AAPL", "GOOGL","SAAB-B.ST"))
}
\references{
Source : https://query2.finance.yahoo.com/v8/finance/chart/
}
