% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_data.R
\name{get_yahoo_data}
\alias{get_yahoo_data}
\title{Get Historical Financial Data For a Given Symbol}
\usage{
get_yahoo_data(
  symbol = "AAPL",
  start_date = NULL,
  end_date = NULL,
  range = "1d",
  .verbose = T
)
}
\arguments{
\item{symbol}{\code{character} - A character string representing the symbol for the financial instrument (e.g., "AAPL" for Apple).}

\item{start_date}{\code{character} - A character string representing the start date in R \code{date} format (UTC time). If \code{NULL}, data starts from 1970-01-01.}

\item{end_date}{\code{character} - A character string representing the end date in a valid R \code{date} format (UTC time). If \code{NULL}, data is retrieved up to the current date.}

\item{range}{\code{character} - A character string representing the range for the returned datas : default will return daily values.\tabular{l}{
   \strong{Valid \code{range} values} \cr
   \code{"1d"}, \code{"5d"}, \code{"1mo"}, \code{"3mo"}, \code{"6mo"}, \code{"1y"}, \code{"2y"}, \code{"5y"}, \code{"10y"}, \code{"ytd"} and \code{"max"} \cr
}}

\item{.verbose}{\code{logical} If TRUE, messages are displayed, e.g., when invalid symbols are detected.}
}
\value{
A \code{data.frame} containing the historical financial data with the following columns:
\item{open}{\code{numeric} The opening price for the period (default is each day).}
\item{close}{\code{numeric} The closing price for the period (default is each day).}
\item{low}{\code{numeric} The lowest price for the period (default is each day).}
\item{high}{\code{numeric} The highest price for the period (default is each day).}
\item{volume}{\code{integer} The traded volume.}
\item{timestamp}{\code{integer} Unix timestamps corresponding to each data point.}
\item{date}{\code{POSIXct} The day of the financial data point.}
\item{currency}{\code{character} The currency in which the data is reported, depending on the marketplace.}
\item{symbol}{\code{character} The stock or financial instrument symbol (e.g., "AAPL").}
\item{shortname}{\code{character} The abbreviated name of the company or financial instrument.}
\item{longname}{\code{character} The full name of the company or financial instrument.}
\item{exchangename}{\code{character} The name of the exchange marketplace where the financial instrument is listed.}
\item{fullexchangename}{\code{character} The full name of the exchange marketplace.}
\item{timezone}{\code{character} The timezone in which the data is reported.}
\item{gmtoffset}{\code{integer} The UNIX timestamp of difference between the market local time and the GMT time.}
\item{regularMarketPrice}{\code{numeric} The actual price if market is open, or last closing price if not.}
\item{fiftyTwoWeekLow}{\code{numeric}  Lowest price for the last 52 weeks.}
\item{fiftyTwoWeekHigh}{\code{numeric} Highest price for the last 52 weeks.}
\item{regularMarketDayHigh}{\code{numeric} The highest price of the day (local exchange place day).}
\item{regularMarketDayLow}{\code{numeric}  The lowest price of the day (local exchange place day).}
}
\description{
Get historical financial values associated with a ticker symbol.
The data includes open stock price, high, low, close, volume, along with timestamps.
}
\details{
The default \code{data.frame} have a line for each day.
If the user provide another range than '1d' (one day), lines will be filtered out, in order to match the desired range.
Valid ranges are: "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y", "ytd", and "max".

The function allows the user to specify a date range using start and end dates. If no date range is specified,
it retrieves all available data from the beginning of time (default for start) to the current date (default for end).
}
\examples{
\dontrun{
data <- get_yahoo_data(symbol = "SAAB-B.ST", start_date = "2020-01-01", range = "1d")
}
}
\references{
Source : https://query1.finance.yahoo.com/v8/finance/chart/
}
\keyword{internal}
