% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_fn.R
\name{last_crypto}
\alias{last_crypto}
\title{Get Latest Cryptocurrencies Values (USD) and Market Data}
\usage{
last_crypto(keep = NULL, .verbose = T)
}
\arguments{
\item{keep}{A character vector of symbols to filter the results. If NULL (default),
no filtering is applied, and all available indices are returned.}

\item{.verbose}{\code{logical}, default = \code{TRUE}. If \code{TRUE}, send messages to the console if necessary.}
}
\value{
A data frame with 100 observations and 13 variables:
\describe{
\item{symbol}{\code{character} - Cryptocurrency ticker symbol.}
\item{name}{\code{character} - Name of the cryptocurrency, along with its quoted 'real-world' currency.}
\item{price}{\code{numeric} - Current price of the cryptocurrency in USD ($).}
\item{change}{\code{numeric} - Absolute price change in USD since the last closing of the exchange place.}
\item{change_percent}{\code{numeric} - Percentage price change since the last closing of the exchange place (\%).}
\item{market_cap}{\code{numeric} - Total market capitalization of the cryptocurrency.}
\item{volume}{\code{numeric} - 24-hour trading volume.}
\item{volume_in_currency_24hr}{\code{numeric} - 24-hour trading volume in the associated 'real-world' currency (most of the time redundant with the \code{volume} column).}
\item{total_volume_all_currencies_24hr}{\code{numeric} - 24-hour total trading volume across all currency pairs.}
\item{circulating_supply}{\code{numeric} - Total circulating supply of the cryptocurrency.}
\item{X52_wk_change_percent}{\code{numeric} - Percentage change in price over the last 52 weeks (\%).}
\item{from}{\code{character}, the currency converted into another, e.g., if the \code{from} value is 1$ ('USD'), you want to receive a certain amount of the other currency to reach 1$.}
\item{to}{\code{character}, the currency that you want to convert into : \strong{all the \code{numeric} values (not \code{integer}) in this line of the \code{data.frame} are expressed with this currency}.}
}
}
\description{
Return a \code{data.frame} of latest financial data for 100 cryptocurrencies, e.g., actual values of each cryptocurrency in USD.
}
\details{
Return a \code{data.frame} with additionnal attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
}
}
\examples{
krypto <- last_crypto()
head(krypto)
}
\references{
Source : Yahoo Finance 'crypto' page, \url{https://finance.yahoo.com/markets/crypto/all/}
}
\seealso{
For more details on the 'last_' family of functions see the help vignette:
\code{vignette("last_family")}
}
