% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_fn.R
\name{last_currencies}
\alias{last_currencies}
\title{Get Currencies Names, Symbols and Add Latest 'USD' Values}
\usage{
last_currencies(keep = NULL, get_changes = F, .verbose = T)
}
\arguments{
\item{keep}{\code{character} - Character vector of symbols to filter the results (perl expression, ignoring case). If \code{NULL} (default),
no filtering is applied, and all available indices are returned.}

\item{get_changes}{\code{logical}, default = \code{FALSE} - If \code{TRUE}, add latest market values ('USD') from \code{get_changes()}. See \code{\link[=get_changes]{get_changes()}} for the base structure.}

\item{.verbose}{\code{logical}, default = \code{TRUE} - If \code{TRUE}, send messages to the console if necessary.}
}
\value{
A \code{data.frame} containing unique financial indices (currencies). The table has
columns like \code{symbol}, \code{name}, and other relevant information, with all column names in lowercase.
If \code{keep} is specified, only the matching currencies are returned. If \code{get_changes} is specified, last financial insights from \code{get_changes()} will be added, e.g., last market values of each currency ('USD')
}
\description{
Return a \code{data.frame} of financial indices (currencies) given a vector of symbols, e.g., 'EUR'.
Optionally, it can filter the results and add the last market price ('USD') of these currencies.
}
\details{
Return a \code{data.frame} with additional attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date.fetch}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
\item{date.begin}{\code{POSIXct} - The oldest observation in the dataset.}
\item{date.end}{\code{POSIXct} - The most recent observation in the dataset.}
\item{date.dif}{\code{numeric} - The number of \emph{seconds} between date.begin and date.end, equivalent of \code{difftime} value.}
}
See \code{vignette("Functions_summary", package = "financr")}
}
\examples{
# Fetch all available indices
all_indices <- last_currencies()

# Fetch only specific indices
selected_indices <- last_currencies(keep = c("^Z", "EUR"))
}
\references{
Source : https://query2.finance.yahoo.com/v1/finance/currencies
}
\seealso{
\code{\link{get_changes}}

For more details see the help vignette:
\code{vignette("currencies", package = "financr")}
}
