% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_summary.R
\name{market_summary}
\alias{market_summary}
\title{Retrieve Market Index Summary Latest Data}
\usage{
market_summary(region = NULL, .verbose = TRUE)
}
\arguments{
\item{region}{\code{character}, default = \code{NULL} (no filtering) - Select a region for filter out the results.\tabular{l}{
   \strong{Valid \code{region} values} \cr
   \code{"US"} , \code{"AU"} , \code{"CA"} , \code{"FR"} , \code{"DE"} , \code{"HK"} , \code{"US"} , \code{"IT"} , \code{"ES"} , \code{"GB"} , \code{"IN"} \cr
}}

\item{.verbose}{\code{logical}, default = \code{FALSE} - Logical flag indicating whether to print verbose output for debugging or informational purposes.}
}
\value{
A data frame with 15 rows and 38 variables. Each row corresponds to a market index and includes the following information:
\itemize{
\item \code{symbol}, \code{shortname}, \code{longname}, \code{exchange}, \code{region}, \code{currency}
\item Market time info: \code{regularmarkettime_raw}, \code{regularmarkettime_fmt}
\item Market pricing info: \code{regularmarketprice_raw}, \code{regularmarketchange_raw}, \code{regularmarketchangepercent_raw}, etc.
\item Metadata such as \code{marketstate}, \code{quotetype}, \code{pricehint}, \code{exchangedatadelayedby}, \code{hasprepostmarketdata}
\item Other fields such as \code{cryptotradeable}, \code{tradeable}, \code{triggerable}, \code{contracts}, etc.
}
}
\description{
Return a \code{data.frame} with latest summary information for a set of major financial market indices.
}
\examples{
df <- market_summary()

df_fr <- market_summary(region = "FR")
}
