% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_assets.R
\name{search_asset_quick}
\alias{search_asset_quick}
\title{Search Financial Asset From a Keyword (Single Text)}
\usage{
search_asset_quick(
  texts = "Dow Jones",
  .verbose = TRUE,
  region = NULL,
  lang = "en"
)
}
\arguments{
\item{texts}{\code{character} string representing the search texts. This can be a company name, index, or financial term. Default is "Dow Jones".}

\item{.verbose}{\code{logical} If \code{TRUE}, print additional details about the search process. Default is \code{TRUE}.}

\item{region}{A character string specifying the region for the search (e.g., "US", "EU"). Default is \code{NULL}, meaning no region filter is applied.}

\item{lang}{A character string specifying the language of the data returned. Default is "en" for English.}
}
\value{
A \code{data.frame} with assets symbols and names.
\itemize{
\item \code{symbol}: The financial symbol (e.g., stock ticker or index).
\item \code{name}: The full name of the financial entity (e.g., 'Dow Jones Industrial Average').
\item \code{exch}: The exchange on which the symbol is listed (e.g., 'DJI', 'CBT').
\item \code{type}: The type of the financial instrument (e.g., I for Index, F for Futures, E for Exchange Traded Fund).
\item \code{exchdisp}: The exchange name displayed (e.g., 'Dow Jones').
\item \code{typedisp}: A long name for the type of financial instrument (e.g., 'Index', 'Futures' or 'ETF' for 'Exchange Traded Fund').
}
}
\description{
Searches for financial data based on a free-text texts and retrieve asset symbol(s), name(s), exchange, and asset type.
}
\examples{
\dontrun{
# Example of searching for financial data related to "Dow Jones"
results <- search_asset_quick(texts = "Dow Jones")
}
}
