% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_text.R
\name{search_from_text}
\alias{search_from_text}
\title{Get Latest Financial Information and Ticker Symbol From Company Name(s)}
\usage{
search_from_text(texts, .verbose = F, exchange = NULL, quotetype = NULL)
}
\arguments{
\item{texts}{\code{character} A character string or a list of character representing the text(s) to search for, e.g., company names.}

\item{.verbose}{\code{logical}, default = \code{FALSE} If TRUE, messages are displayed, e.g., when invalid symbols are detected.}

\item{exchange}{(optionnal)  A character string representing the exchange place(s) to consider - exact match, e.g., 'STO' (Stockholm stock exchange). Default keep all results.}

\item{quotetype}{(optionnal) A character string representing the type of entry to search for, e.g., a stock market ('ETF', 'EQUITY' or 'MUTUALFUND') or a major indice ('INDEX').}
}
\value{
A data frame with the following columns:
\item{symbol}{\code{numeric} - The ticker symbol associated with an entry.}
\item{exchange}{\code{character} - .}
\item{shortname}{\code{character} - .}
\item{quotetype}{\code{numeric} - .}
\item{index}{\code{character} - .}
\item{score}{\code{numeric} - .}
\item{typedisp}{\code{character} - .}
\item{longname}{\code{character} - .}
\item{exchdisp}{\code{character} - .}
\item{isyahoofinance}{\code{character} - .}
\item{searched}{\code{character} - .}
}
\description{
Get latest overall insights and ticker symbols associated with free texts, such as companies names.
Several ticker symbols associated with various exchanges places could be returned, with
, companies names, last price on the marketplace,
sector/category (if available), etc.
}
\details{
Return a \code{data.frame} with additionnal attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
}
}
\examples{
oil <- search_from_text(texts = c("TOTAL", "SHELL", "BP"),  quotetype = "EQUITY")
#Get data on marketplace(s)
swedish <- search_from_text(c("SAAB", "VOLVO"),  exchange = c("STO", "PAR"))
}
\references{
Source : https://query2.finance.yahoo.com/v1/
}
