% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_assets.R
\name{search_summary}
\alias{search_summary}
\title{Get Latest Financial Information and Ticker Symbol From Company Name(s)}
\usage{
search_summary(texts, exchange = NULL, sector = NULL)
}
\arguments{
\item{texts}{A character string representing the company name to search for.}

\item{exchange}{(optionnal) A character string representing the exchange place(s) to consider (exact match). Default keep all the exchange places.}

\item{sector}{(optionnal)  A character string representing the sector(s) to consider (exact match). Default keep all results.}
}
\value{
A data frame with columns:

\describe{
\item{symbol}{\code{character} - The ticker symbol associated with an asset, e.g., "VTI", "^DWCPF".}
\item{name}{\code{character} - Name of the asset, e.g., ETF or index name.}
\item{last_price}{\code{numeric} - The last price of the asset.}
\item{sector}{\code{character} - The sector or industry category if available, e.g., 'Industrials', 'Consumer Cyclical'.}
\item{type}{\code{character} - The type of asset (certainly "stocks").}
\item{exchange}{\code{character} - The stock exchange place for this asset, e.g., 'PAR' is Paris (FR) exchange place.}
\item{searched}{\code{character} - The text searched on Yahoo, e.g., "Dow Jones".}
}
}
\description{
Get latest overall insights and ticker symbols associated with free texts, such as companies texts.
Several ticker symbols associated with various exchanges places could be returned, with
, companies texts, last price on the marketplace,
sector/category (if available), etc.
}
\details{
Return a \code{data.frame} with additionnal attributes:
\describe{
\item{n.currencies}{\code{integer} - Number of unique currencies in the \code{data.frame}.}
\item{currencies}{\code{character} - A vector of currency symbols in the \code{data.frame} (e.g., \code{"BTC"}, \code{"ETH"}, \code{"USD"}).}
\item{exchange}{\code{character} - A vector of exchange pairs (e.g., \code{"BTC => USD"}). If not applicable, \code{NULL} (no entry).}
\item{date}{\code{Date} - The date when the data was retrieved, set using \code{Sys.Date()}.}
\item{crypto}{\code{logical} - \code{TRUE} if cryptocurrencies are present, otherwise \code{FALSE}.}
}
}
\examples{
oil <- search_summary(texts = c("Dow Jones", "Euronext"))
#Get data on marketplace(s)
cars <- search_summary(texts = c("RENAULT", "VOLVO"),  exchange = c("STO", "PAR"))
}
\references{
Source : search on https://finance.yahoo.com/lookup/
}
\seealso{
For more details see the help vignette:
\code{vignette("get_info_and_historic", package = "financr")}
}
