% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_and_similar_symbol.R
\name{similar_symbol}
\alias{similar_symbol}
\title{Find Similar Financial Symbols}
\usage{
similar_symbol(symbols = "SAAB-B.ST", .verbose = F, ...)
}
\arguments{
\item{symbols}{\code{character} A character string representing a financial symbol to search.}

\item{.verbose}{\code{logical} If TRUE, messages are displayed when invalid symbols are detected. Default is TRUE.}

\item{...}{Other symbols (char. or list of char.)}
}
\value{
A boolean table with one row and as many columns as the number of \emph{unique} symbols provided by the user.
Each column corresponds to a symbol, with TRUE if Yahoo Finance recognizes the symbol, and FALSE otherwise.
}
\description{
Given symbol(s), retrieve identical symbols (according to Yahoo Finance) and a score of similarity.
}
\examples{
similar_symbol(symbols =   "AAPL,GOOGL")
similar_symbol(symbols =   c("AAPL", "GOOGL"))
}
\references{
Source : https://query2.finance.yahoo.com/v6/finance/quote/validate
}
