% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_df.R
\name{standardize_df_split_col_w_several_values}
\alias{standardize_df_split_col_w_several_values}
\title{Split df cols according to a separator (space the default)}
\usage{
standardize_df_split_col_w_several_values(
  df,
  separator = " ",
  col_tolerate_separator = "name|symbol|date|time|exch"
)
}
\arguments{
\item{df}{\code{data.frame} of raw values}

\item{separator}{\code{character} of safe separator for spliting values}

\item{col_tolerate_separator}{\code{character} of regex for colname that DON'T need to be split}
}
\description{
Split df cols according to a separator (space the default)
}
