% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_symbol.R
\name{valid_symbol}
\alias{valid_symbol}
\title{Validate Financial Symbols}
\usage{
valid_symbol(symbols = NULL, ..., .verbose = T)
}
\arguments{
\item{symbols}{\code{character} A string or a list of character strings representing financial symbols to validate.}

\item{...}{Other symbols (char. or list of char.)}

\item{.verbose}{\code{logical} If TRUE, messages are displayed when invalid symbols are detected. Default is TRUE.}
}
\value{
A \code{data.frame} boolean \code{logical} table with one row and as many columns as the number of unique symbols provided by the user.
Each column corresponds to a symbol, and the \code{logical} value is \code{TRUE} if Yahoo Finance recognizes the symbol, and \code{FALSE} otherwise.
}
\description{
Checks the validity of financial symbols using Yahoo Finance's validation API.
Return a \code{data.frame} of \code{logical} values indicating whether each symbol is recognized by Yahoo Finance.
}
\examples{
valid_symbol("AAPL,GOOGL")
valid_symbol(symbols = c("CDF", "SCR", "INVALID"))
}
\references{
Source : https://query2.finance.yahoo.com/v6/finance/quote/validate
}
