% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_changes_historic.R
\name{get_changes_historic}
\alias{get_changes_historic}
\title{Get Historic of Exchange Rates For Devises}
\usage{
get_changes_historic(
  from = NULL,
  to = "USD",
  interval = "1d",
  range = "1y",
  .verbose = T
)
}
\arguments{
\item{from}{default = \code{NULL} A character string representing the base currency (e.g., "USD").
The user must provide \code{from} values with one of the following method :
\itemize{
\item passing a paired character list, e.g., c("EUR" = "USD", "RON" = "EUR") ;
\item or passing one or several \code{from} values :
\itemize{
\item associated with a single \code{to} value, e.g., \verb{from = c('EUR', "RON"), to = "USD"}
\item or associated with a list of \code{from} values of the same length, e.g., \verb{from = c('EUR', "RON"), to = c("USD", "EUR")} .
}
}}

\item{to}{default = \code{"USD"} A character string representing the target currencies, e.g., \code{c("EUR", "USD")}.}

\item{interval}{\code{character}, default = \code{"1d"}. The interval between 2 rows of the time.series answered\tabular{l}{
   \strong{Valid \code{interval} values} \cr
   \code{"1m"}, \code{"2m"}, \code{"3m"}, \code{"5m"}, \code{"15m"}, \code{"30m"}, \code{"1h"}, \code{"4h"}, \code{"1d"}, \code{"1wk"}, \code{"1mo"}, \code{"3mo"} \cr
}}

\item{range}{\code{character}, default = \code{"1y"}. The period covered by the time series.\tabular{l}{
   \strong{Valid \code{range} values} \cr
   \code{"1d"}, \code{"5d"}, \code{"1m"}, \code{"3m"}, \code{"6m"}, \code{"1y"}, \code{"5y"}, \code{"ytd"}, \code{"all"} \cr
}}

\item{.verbose}{\code{logical} If TRUE, messages are displayed, e.g., when invalid symbols are detected. Default is FALSE.}
}
\value{
A \code{data.frame} containing the historical exchanges rates on a given period (default is daily results for a year).
The columns are:
\item{timestamp}{\code{POSIXct} The opening price for the period (default is each day).}
\item{close}{\code{numeric} The closing price for the period (default is each day).}
\item{high}{\code{numeric} The lowest price for the period (default is each day).}
\item{low}{\code{numeric} The highest price for the period (default is each day).}
\item{open}{\code{integer} The traded volume.}
\item{from}{\code{character}, the currency converted into another, e.g., if the \code{from} value is 1$ ('USD'), you want to receive a certain amount of the other currency to reach 1$.}
\item{to}{\code{character}, the currency that you want to convert into : \strong{all the \code{numeric} values (not \code{integer}) in this line of the \code{data.frame} are expressed with this currency}.}
Depending on the desired interval, recent observation will be truncated,
e.g., a \code{'5y'} range  with a \code{'1d'} interval will answer approximately 30 days of values from 5 years ago.
}
\description{
Get a \code{data.frame} of historical values of exchanges rates, given currencies to exchange.
Default parameters will return a period of 1 year (1 obs. per day) and convert to USD.
}
\details{
For each pair of 'from' and 'to' currency, it returns a \code{data.frame} with the historical exchanges rates on a given period (default is daily results for a year)
}
\examples{
# 1. Define currencies to exchange with the syntax of get_changes().
# Default convert to 'USD"
# e.g., get exchange historical data of exchanges rates from € and ¥ to $ :
df_days <- get_changes_historic(from = c("EUR", "JPY"))
# Or pass a named list of character
df2_days <- get_changes_historic(from = c("EUR" = "RON", "USD" = "EUR"))
# Or pass paired values as 2 list (equivalent to hereabove line) :
same_as_df2_days <- get_changes(from = c("EUR", "USD"), to =c("RON" , "EUR"))
# Default answer is a `data.frame` with  1 obs. per day for 1 year

# 2. Tweak the time.series with the parameters interval and range :
df_month <- get_changes_historic(from = c("EUR", "JPY"), interval = "1mo", range = '5y')
# It's 1 obs. per month for a 5-years period
}
\references{
Source : https://query1.finance.yahoo.com/v8/finance/chart/
}
\seealso{
\code{\link{get_changes}}
}
