% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.R
\name{get_historic_light}
\alias{get_historic_light}
\title{Get Historical Financial Data For Ticker Symbols}
\usage{
get_historic_light(
  symbols = "SAAB-B.ST",
  interval = "1d",
  range = "1mo",
  .verbose = F
)
}
\arguments{
\item{symbols}{\code{character} A character string representing the financial indices to search for, e.g., ticker symbol(s).}

\item{interval}{\code{character}, default = \code{"1d"}. The interval between 2 rows of the time.series answered\tabular{l}{
   \strong{Valid \code{interval} values} \cr
   \code{"1m"},  \code{"5m"}, \code{"15m"}, \code{"1d"},  \code{"1wk"}, \code{"1mo"} \cr
}}

\item{range}{\code{character}, default = \code{"1y"}. The period covered by the time series.\tabular{l}{
   \strong{Valid \code{range} values} \cr
   \code{"1d"}, \code{"5d"}, \code{"1mo"}, \code{"3mo"}, \code{"6mo"}, \code{"1y"}, \code{"5y"}, \code{"max"} \cr
}}

\item{.verbose}{\code{logical}, default = \code{TRUE}. If \code{TRUE}, send messages to the console.}
}
\value{
A \code{data.frame} with a historical values :
\describe{
\item{symbol}{\code{character} - Financial ticker symbol.}
\item{timestamp}{\code{POSIXct} - Date of the observation (closing price).}
\item{close}{\code{numeric} - Closing price of the asset.}
}
}
\description{
Get an historic of stock market data for financial ticker symbols, e.g., values at each closing day.
}
\examples{
histo_light <- get_historic_light(c("SAAB-B.ST", "AAPL"))
}
